feat: 여행지 1개 상세조회 API 구현 및 imageUrl이 없을 경우 처리

+++ Tlog/app/src/main/java/com/tlog/ui/component/travel/review/ReviewSection.kt
@@ -27,7 +27,7 @@ import com.tlog.ui.theme.MainFont
 
 @Composable
 fun ReviewSection(
-    avgStarRating: Float,
+    avgStarRating: Double,
     ratingDistribution: Map<String, Int>,
     reviewList: List<DetailReview>,
     reviewCnt: Int = Int.MAX_VALUE
